Additional and Resampled Image Training
Characters who are getting bad results actually happen to be the ones who we have a lot of training images for (e.g. Homer), however they are not being used in the training process as other characters have less.
This commit focuses on increasing the number of training images.
Steps were:

1. Made GoogleImageDownloder.py to bulk download 100 images at a time (via a GITHUB clone and modify)

2. Used this code to bulk rename files in Bash, where 'a' is the starting image number, so they meet the required naming conventions:

a=1342
for i in *.jpg; do
  new=$(printf "pic_%d.jpg" "$a")
  mv -i -- "$i" "$new"
  let a=a+1
done

3. Manually filtered out bad images

4. Used this Bash code to flip images for resampling for where we still needed more (e.g. Milhouse who is hard to get enough photos of online)

for szFile in *.jpg
do
    convert "$szFile" -flop "$(basename "$szFile")" ;
done

This has beefed up training numbers significantly however will need to see if it actually improves the result.